pyproject.toml
[project]
name = "congestion-pricing-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.4.3",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "playwright>=1.52.0",
    "poetry>=2.1.3",
    "praw>=7.8.1",
    "pytest-playwright>=0.7.0",
    "python-dotenv>=1.1.0",
]
